fix: extract shared highlightMarkers, fix stale closure in onEditSave
- Extract highlightMarkers + sanitize to src/utils/transcript-markdown.ts
- TranscriptBrief and TranscriptEditor now use shared renderBriefHtml()
- Fix onEditSave stale closure: use functional setBriefResult updater
